Lots of good stuff...
So as it turns out, many light-weight LLMs do not use a System Prompt.
Lesson learned. Moving tagging prompts to Human Prompt.

Also, discovered a silly bug where I was stopping on "\n\n" from the
pre-processor. Not sure how all those on-stop commands got there. But
removing them to allow the LLM to provide its own stop... way better.

Continuing to make the overall experience a with the UI a good one.