Implement ParentDocumentRetriever functions
Devise a method for retrieving a single large response
based on smaller child matches. This is more or less an
experiment. I am not exactly sure how much this bring
to the table for a chat-based system.

I have yet to test this with pdf importer. It may truly
benefit in that regard.

Squeezing every last token of importance to the context
window, has been my focus. I'd like to next figure out
how to remove any similar smaller children matches, now
that the larger full-document match is being thrown into
the context window...