Enable thinking animation when tagging
Allow the pulse animation to fire off again when the LLM is doing
something other than responding. This is obviously a per-LLM thing.
The largest (smarter) the model, the easier it is to train it how
to create metadata tags.

Fixing typos in system prompt. I need to spend more time with the
other models to polish those up.

Adjusted the light-mode colors slightly to be more contrastive.

Doubled the default minimum context window size. 2048 is practically
the size of the system prompts alone...