Adjust heat map, drop story telling...
De-duplication shouldn't really have a color. I am not even sure
if less or more is even better?

I think I'd like to see red if there is an excessive amount when
compared to a small amount of context. So what I should do is is
focus on context vs dedups instead of total de-dups retrieved
from RAG.