Parse meta_tags more robustly
Trying to make the capture and hidding of <meta_tags as robust as
possible. The LLM Mixtral is a PITA.

Beautify the response window a little bit (pulsating name, but not
so distracting).

Adjust prompts to have plural entries where they should (helps the
LLM understand this is a list, this is a string, etc)