Move tagging to pre-processor only
No longer ask the heavy weight LLM to provide a story, *and* tag
what it said. Leave that entirely up to the light weight
pre-processors. This helps tremendously with knocking down the times
the heavy-weight LLM botches metadata formatting. And truly frees
up the token count to boot.

Because of the way we want to discovery entities in a user's query,
it is necessary to run the pre-processor twice (once to discover
entities, then again with entities populating the context window
(character sheets) possibly existing character sheet included. This
adds a little extra time for each turn, but because we are doing a
back-to-back request, I have found Ollama is very quick at inference.

Added aside comments your LLM assistant can chime in on, to make it
feel more like a DM.

Modify all the prompts to be more Forgotten Realms-ish. Many LLMs
know plenty about this campaign setting, so its a bit easier to use
for early adoption.

Add ability to use multiple heavy-weight LLMs. When the pre-processor
detects NSFW content, and there is a NSFW available model to use, it
will use that instead of the normal model. I ran into odd issues
when violence ensued during the story, some models will refuse to
generate content... even in a fantasy trope world. Sheesh.

Fixed a bug when reaching the same amount of turns as the user allowed
chat_histories (history_sessions), would cause a divide by zero...

Added a new Scene Manager class to begin helping with better
transitional data between turns.